requirements.txt
flask==3.0.3
gunicorn==22.0.0
requests==2.32.3
python-telegram-bot==21.6
pytz==2024.1
